+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@ Release 3.4
 
 * #421 Per-request EncoderFilters
 * #421 Custom per-request Encoder is ignored on update
+* #423 Wrong Encoder for target entities in overlayed relationships
 
 Release 3.3